Add buffer reference to Screen, editor, and change code for clarity
The screen should have an optional reference to the buffer it should
display. The buffer is optional, because if there is no buffer, it
should show a welcome message that will be added later. The editor
struct should also have a reference to this buffer. Later, that will
change so the editor struct holds a Vector of Buffer structs, and
scren will only hold a reference to a single buffer.

Some variables and structs were also renamed for better clarity.