Delete file_props.rs for now and have the buffer keep track of line endings
file_props.rs was only doing one thing, and that's keep track of line
endings. This had a side effect of not allowing the user to delete
characters when no file was provided, as file_props was None. The
buffer should keep track of the line endings anyway, what if the user
wants to use use a temporary buffer not associated with a file?