Refactor the syncing of the ropey and terminal cursors and insert tabs
This commit refactors the syncing of the ropey and terminal cursors by
putting the loginc into a fuction in the buffer.rs file.
It also inserts tabs when the tab key is pressed, but they are assumed
to be of a character width of 8, so that might brake a few terminals.